feat: unificar entrada CLI
